@@ -1,41 +1,36 @@
 <?php
 
 /**
- * This file is part of the PsuwCommonListener package.
+ * This file is part of the PsuwExpressionEvaluator package.
  *
  * @copyright Copyright (c) 2016 Paweł Suwiński
- * @author Paweł Suwiński <psuw@wp.pl>
  * @license MIT
  */
 
-namespace Psuw\CommonListener\Tests\Expression;
+namespace Psuw\ExpressionEvaluator\Tests\Expression;
 
+use Psuw\ExpressionEvaluator\Expression\ExpressionLanguageAwareInterface;
 use Symfony\Component\ExpressionLanguage\Expression;
-use \Psuw\CommonListener\Expression\ExpressionLanguageAwareInterface;
-
 
 /**
- * ExpressionEvaluatingTraitTest 
- * 
- * @package PsuwCommonListener
+ * ExpressionEvaluatingTraitTest.
+ *
  * @copyright Copyright (c) 2016, Paweł Suwiński
- * @author Paweł Suwiński <psuw@wp.pl> 
  * @license MIT
  */
-class ExpressionEvaluatingTraitTest extends \PHPUnit\Framework\TestCase 
-    implements ExpressionLanguageAwareInterface
+class ExpressionEvaluatingTraitTest extends \PHPUnit\Framework\TestCase implements ExpressionLanguageAwareInterface
 {
-    use \Psuw\CommonListener\Expression\ExpressionEvaluatingTrait;
-    use \Psuw\CommonListener\Expression\ExpressionLanguageAwareTrait;
+    use \Psuw\ExpressionEvaluator\Expression\ExpressionEvaluatingTrait;
+    use \Psuw\ExpressionEvaluator\Expression\ExpressionLanguageAwareTrait;
 
     /**
-     * testAddingExpressionValidationException 
-     * 
+     * testAddingExpressionValidationException.
+     *
      * @dataProvider providerInvalidExpression
      * @expectedException \InvalidArgumentException
      * @expectedExceptionMessageRegExp #string or.*Expression.*expected#
-     * @param mixed $expression 
-     * @return void
+     *
+     * @param mixed $expression
      */
     public function testAddingExpressionValidationException($expression)
     {
@@ -43,8 +38,8 @@ class ExpressionEvaluatingTraitTest extends \PHPUnit\Framework\TestCase
     }
 
     /**
-     * providerInvalidExpression 
-     * 
+     * providerInvalidExpression.
+     *
      * @return array
      */
     public function providerInvalidExpression()
@@ -53,13 +48,13 @@ class ExpressionEvaluatingTraitTest extends \PHPUnit\Framework\TestCase
     }
 
     /**
-     * testSettingExpressionsValidationException 
-     * 
+     * testSettingExpressionsValidationException.
+     *
      * @dataProvider providerSettingExpressionsValidationException
      * @expectedException \InvalidArgumentException
      * @expectedExceptionMessageRegExp #string or.*Expression.*expected#
-     * @param array $expression 
-     * @return void
+     *
+     * @param array $expressions
      */
     public function testSettingExpressionsValidationException($expressions)
     {
@@ -67,125 +62,112 @@ class ExpressionEvaluatingTraitTest extends \PHPUnit\Framework\TestCase
     }
 
     /**
-     * providerSettingExpressionsValidationException 
-     * 
+     * providerSettingExpressionsValidationException.
+     *
      * @return array
      */
     public function providerSettingExpressionsValidationException()
     {
-        return array(
-            [array('event.getSubject()', true, new Expression('true'))],
-            [array(new Expression('event.getSubject()'), false)],
-            [array('event.getSubject()', new \ArrayObject(), 'true')],
-        );
+        return [
+            [['event.getSubject()', true, new Expression('true')]],
+            [[new Expression('event.getSubject()'), false]],
+            [['event.getSubject()', new \ArrayObject(), 'true']],
+        ];
     }
 
     /**
-     * testExpressionSetting 
-     * 
-     * @return void
+     * testExpressionSetting.
      */
     public function testExpressionSetting()
     {
         $this->assertEmpty($this->expressions);
 
         $this->addExpression('true');
-        $this->assertEquals(array('true'), $this->expressions);
+        $this->assertEquals(['true'], $this->expressions);
 
         $this->addExpression('false');
-        $this->assertEquals(array('true', 'false'), $this->expressions);
+        $this->assertEquals(['true', 'false'], $this->expressions);
 
-        $expressions = array('true', 'true');
+        $expressions = ['true', 'true'];
         $this->setExpressions($expressions);
         $this->assertEquals($expressions, $this->expressions);
 
         $this->addExpression('false');
-        $this->assertEquals(
-            array_merge($expressions, array('false')),
-            $this->expressions
-        );
+        $this->assertEquals(array_merge($expressions, ['false']), $this->expressions);
 
-        $this->setExpressions(array());
+        $this->setExpressions([]);
         $this->assertEmpty($this->expressions);
     }
 
     /**
-     * testEvaluatingExpressionInterfaceException 
-     * 
+     * testEvaluatingExpressionInterfaceException.
+     *
      * @expectedException \InvalidArgumentException
      * @expectedExceptionMessageRegExp #ExpressionLanguageAwareInterface#
-     * @return void
      */
     public function testEvaluatingExpressionInterfaceException()
     {
         $mock = $this->getMockForTrait(
-            'Psuw\CommonListener\Expression\ExpressionEvaluatingTrait',
-            array(),
+            'Psuw\ExpressionEvaluator\Expression\ExpressionEvaluatingTrait',
+            [],
             '',
             true,
             true,
             true,
-            array('__invoke')
+            ['__invoke']
         );
         $mock
             ->method('__invoke')
             ->willReturnCallback(\Closure::bind(
-                function(array $context) {
+                function (array $context) {
                     return $this->evaluateExpressions($context);
                 },
                 $mock,
                 $mock
             ));
-        $mock(array());
-    }
-    
-    /**
-     * testEvaluatingEmptyExpressions 
-     * 
-     * @return void
-     */
-    public function testEvaluatingEmptyExpressions()
-    {
-        $this->setExpressions(array());
-        $this->assertEquals(array(), $this->evaluateExpressions(array()));
+        $mock([]);
     }
 
     /**
-     * testEvaluatingExpressions 
-     * 
-     * @return void
+     * testEvaluatingEmptyExpressions.
+     */
+    public function testEvaluatingEmptyExpressions()
+    {
+        $this->setExpressions([]);
+        $this->assertEquals([], $this->evaluateExpressions([]));
+    }
+
+    /**
+     * testEvaluatingExpressions.
      */
     public function testEvaluatingExpressions()
     {
-        $context = array(
-            'container' => new \ArrayObject(array_flip(array('one', 'two'))),
+        $context = [
+            'container' => new \ArrayObject(array_flip(['one', 'two'])),
             'trigger' => true,
-        );
+        ];
 
         $testExpression = 'result === true ? "OK" : "NOT"';
 
         /**
-         * expression => expected value 
+         * expression => expected value.
          */
-        $expressions = array(
+        $expressions = [
             'trigger && container.offsetExists("one")' => true,
             $testExpression => 'OK',
             '!trigger && container.offsetExists("one")' => false,
             $testExpression => 'NOT',
             'trigger && container.offsetExists("five")' => false,
             $testExpression => 'NOT',
-            'trigger && !container.offsetExists("five")' => true, 
+            'trigger && !container.offsetExists("five")' => true,
             $testExpression => 'OK',
-            'container["one"] + container["two"]' => 1, 
+            'container["one"] + container["two"]' => 1,
             'result + container["two"]' => 2,
             'container["one"] * container["two"]' => 0,
-        );
+        ];
 
         $this->setExpressions(array_keys($expressions));
 
-        $this->assertEquals(
-            array_values($expressions),
-            $this->evaluateExpressions($context)
-        );
+        $this->assertEquals(array_values($expressions), $this->evaluateExpressions($context));
     }
 }
