@@ -1,23 +1,20 @@
 <?php
 
 /**
- * This file is part of the PsuwCommonListener package.
+ * This file is part of the PsuwExpressionEvaluator package.
  *
  * @copyright Copyright (c) 2016 Paweł Suwiński
- * @author Paweł Suwiński <psuw@wp.pl>
  * @license MIT
  */
 
-namespace Psuw\CommonListener\Expression;
+namespace Psuw\ExpressionEvaluator\Expression;
 
 use Symfony\Component\ExpressionLanguage\ExpressionLanguage;
 
 /**
- * ExpressionLanguageAwareTrait 
- * 
- * @package PsuwCommonListener
+ * ExpressionLanguageAwareTrait.
+ *
  * @copyright Copyright (c) 2016, Paweł Suwiński
- * @author Paweł Suwiński <psuw@wp.pl> 
  * @license MIT
  */
 trait ExpressionLanguageAwareTrait
@@ -25,10 +22,7 @@ trait ExpressionLanguageAwareTrait
     protected $expressionLanguage;
 
     /**
-     * setExpressionLanguage 
-     * 
-     * @param ExpressionLanguage $expressionLanguage 
-     * @return void
+     * setExpressionLanguage.
      */
     public function setExpressionLanguage(ExpressionLanguage $expressionLanguage)
     {
@@ -36,15 +30,16 @@ trait ExpressionLanguageAwareTrait
     }
 
     /**
-     * getExpressionLanguage 
-     * 
-     * @return ExpressionLanguage 
+     * getExpressionLanguage.
+     *
+     * @return ExpressionLanguage
      */
     public function getExpressionLanguage()
     {
-        if(!isset($this->expressionLanguage)) {
+        if (!isset($this->expressionLanguage)) {
             $this->expressionLanguage = new ExpressionLanguage();
         }
+
         return $this->expressionLanguage;
     }
 }
